+book title preview
The metal cast font is being developed the cover of _Ghost Girl and the Ghost Giant_. As such, a preview of said title has been added.

The preview code was also rewritten in order to have a reference to the compiled SVG to make it easier to drop specific characters into the preview.